LPS-109350 Move default guest group layout creation code to component
